remove the EmailData component
