+++ client/components/data/domain-management/index.jsx
@@ -11,15 +11,15 @@ import { connect } from 'react-redux';
 /**
  * Internal dependencies
  */
-import StoreConnection from 'components/data/store-connection';
 import CartStore from 'lib/cart/store';
-import QueryProductsList from 'components/data/query-products-list';
-import QuerySitePlans from 'components/data/query-site-plans';
-import QuerySiteDomains from 'components/data/query-site-domains';
-import QueryContactDetailsCache from 'components/data/query-contact-details-cache';
+import DnsStore from 'lib/domains/dns/store';
+import EmailForwardingStore from 'lib/domains/email-forwarding/store';
+import { fetchUsers } from 'lib/users/actions';
+import { fetchByDomain, fetchBySiteId } from 'state/google-apps-users/actions';
+import { getByDomain, getBySite, isLoaded } from 'state/google-apps-users/selectors';
+import { getCurrentUser } from 'state/current-user/selectors';
 import { getPlansBySite } from 'state/sites/plans/selectors';
 import { getSelectedSite } from 'state/ui/selectors';
-import PageViewTracker from 'lib/analytics/page-view-tracker';
 import { getDecoratedSiteDomains, isRequestingSiteDomains } from 'state/sites/domains/selectors';
 import { getProductsList } from 'state/products-list/selectors';
 import NameserversStore from 'lib/domains/nameservers/store';
@@ -30,12 +30,15 @@ import {
 	fetchWapiDomainInfo,
 	fetchWhois,
 } from 'lib/upgrades/actions';
+import PageViewTracker from 'lib/analytics/page-view-tracker';
+import QueryContactDetailsCache from 'components/data/query-contact-details-cache';
+import QueryProductsList from 'components/data/query-products-list';
+import QuerySitePlans from 'components/data/query-site-plans';
+import QuerySiteDomains from 'components/data/query-site-domains';
+import SiteRedirectStore from 'lib/domains/site-redirect/store';
+import StoreConnection from 'components/data/store-connection';
 import UsersStore from 'lib/users/store';
 import WapiDomainInfoStore from 'lib/domains/wapi-domain-info/store';
-import { fetchUsers } from 'lib/users/actions';
-import SiteRedirectStore from 'lib/domains/site-redirect/store';
-import DnsStore from 'lib/domains/dns/store';
-import EmailForwardingStore from 'lib/domains/email-forwarding/store';
 import WhoisStore from 'lib/domains/whois/store';
 
 function getStateFromStores( props ) {
@@ -45,6 +48,8 @@ function getStateFromStores( props ) {
 		domains: props.selectedSite ? props.domains : null,
 		dns: DnsStore.getByDomainName( props.selectedDomainName ),
 		emailForwarding: EmailForwardingStore.getByDomainName( props.selectedDomainName ),
+		googleAppsUsers: props.googleAppsUsers,
+		googleAppsUsersLoaded: props.googleAppsUsersLoaded,
 		isRequestingSiteDomains: props.isRequestingSiteDomains,
 		location: SiteRedirectStore.getBySite( props.selectedSite.domain ),
 		nameservers: NameserversStore.getByDomainName( props.selectedDomainName ),
@@ -52,6 +57,7 @@ function getStateFromStores( props ) {
 		selectedDomainName: props.selectedDomainName,
 		selectedSite: props.selectedSite,
 		sitePlans: props.sitePlans,
+		user: props.currentUser,
 		users: UsersStore.getUsers( { siteId: props.selectedSite.ID } ),
 		wapiDomainInfo: WapiDomainInfoStore.getByDomainName( props.selectedDomainName ),
 		whois: WhoisStore.getByDomainName( props.selectedDomainName ),
@@ -71,6 +77,7 @@ class DomainManagementData extends React.Component {
 		needsDomains: PropTypes.bool,
 		needsDomainInfo: PropTypes.bool,
 		needsEmailForwarding: PropTypes.bool,
+		needsGoogleApps: PropTypes.bool,
 		needsNameservers: PropTypes.bool,
 		needsPlans: PropTypes.bool,
 		needsProductsList: PropTypes.bool,
@@ -94,7 +101,7 @@ class DomainManagementData extends React.Component {
 	}
 
 	loadData = prevProps => {
-		const { selectedDomainName, selectedSite } = this.props;
+		const { needsGoogleApps, needsUsers, selectedDomainName, selectedSite } = this.props;
 
 		if ( this.props.needsDns ) {
 			fetchDns( selectedDomainName );
@@ -108,18 +115,26 @@ class DomainManagementData extends React.Component {
 			fetchEmailForwarding( selectedDomainName );
 		}
 
+		if (
+			needsGoogleApps &&
+			( prevProps.needsGoogleApps !== needsGoogleApps || prevProps.selectedSite !== selectedSite )
+		) {
+			this.props.fetchGoogleAppsUsers( selectedSite.ID );
+		}
+
 		if ( this.props.needsNameservers ) {
 			fetchNameservers( selectedDomainName );
 		}
 
-		if ( this.props.needsUsers ) {
-			if ( prevProps.selectedSite !== selectedSite ) {
-				fetchUsers( { siteId: selectedSite.ID, number: 1000 } );
-			}
-		}
-
 		if ( this.props.needsWhois ) {
 			fetchWhois( this.props.selectedDomainName );
+		}
+
+		if (
+			needsUsers &&
+			( prevProps.needsUsers !== needsUsers || prevProps.selectedSite !== selectedSite )
+		) {
+			fetchUsers( { siteId: selectedSite.ID, number: 1000 } );
 		}
 	};
 
@@ -177,8 +192,11 @@ class DomainManagementData extends React.Component {
 				<StoreConnection
 					component={ this.props.component }
 					context={ this.props.context }
+					currentUser={ this.props.currentUser }
 					domains={ this.props.domains }
 					getStateFromStores={ getStateFromStores }
+					googleAppsUsers={ this.props.googleAppsUsers }
+					googleAppsUsersLoaded={ this.props.googleAppsUsersLoaded }
 					isRequestingSiteDomains={ this.props.isRequestingSiteDomains }
 					products={ this.props.productsList }
 					selectedDomainName={ this.props.selectedDomainName }
@@ -191,15 +209,32 @@ class DomainManagementData extends React.Component {
 	}
 }
 
-export default connect( state => {
-	const selectedSite = getSelectedSite( state );
-	const siteId = get( selectedSite, 'ID', null );
+export default connect(
+	( state, { selectedDomainName } ) => {
+		const selectedSite = getSelectedSite( state );
+		const siteId = get( selectedSite, 'ID', null );
+		const googleAppsUsers = selectedDomainName
+			? getByDomain( state, selectedDomainName )
+			: getBySite( state, selectedSite.ID );
 
-	return {
-		domains: getDecoratedSiteDomains( state, siteId ),
-		isRequestingSiteDomains: isRequestingSiteDomains( state, siteId ),
-		productList: getProductsList( state ),
-		sitePlans: getPlansBySite( state, selectedSite ),
-		selectedSite,
-	};
-} )( DomainManagementData );
+		return {
+			currentUser: getCurrentUser( state ),
+			domains: getDecoratedSiteDomains( state, siteId ),
+			googleAppsUsers,
+			googleAppsUsersLoaded: isLoaded( state ),
+			isRequestingSiteDomains: isRequestingSiteDomains( state, siteId ),
+			productsList: getProductsList( state ),
+			sitePlans: getPlansBySite( state, selectedSite ),
+			selectedSite,
+		};
+	},
+	( dispatch, { selectedDomainName } ) => {
+		const googleAppsUsersFetcher = selectedDomainName
+			? () => fetchByDomain( selectedDomainName )
+			: siteId => fetchBySiteId( siteId );
+
+		return {
+			fetchGoogleAppsUsers: siteId => dispatch( googleAppsUsersFetcher( siteId ) ),
+		};
+	}
+)( DomainManagementData );
